requirements.txt
certifi==2020.4.5.1
chardet==3.0.4
cheroot==8.3.0
CherryPy==18.6.0
idna==2.9
importlib-metadata==1.6.0
importlib-resources==1.5.0
jaraco.classes==2.0
jaraco.collections==2.1
jaraco.functools==2.0
jaraco.text==3.2.0
more-itertools==8.2.0
pendulum==2.1.0
portend==2.6
pyTelegramBotAPI==3.7.1
python-dateutil==2.8.1
pytz==2020.1
pytzdata==2019.3
requests==2.23.0
RPi.GPIO==0.7.0
schedule==0.6.0
six==1.14.0
tempora==1.14.1
urllib3==1.25.9
zc.lockfile==2.0
zipp==1.2.0
